重构舆情级别

+++ iCapture/src/com/icapture/service/diy/impl/WarnLevelServiceImpl.java
@@ -10,6 +10,7 @@ import com.connection.db.DBHandle.Base;
 import com.connection.page.Page;
 import com.icapture.entity.diy.WarnLevel;
 import com.icapture.service.diy.WarnLevelService;
+import com.icapture.web.action.CrudEntity;
 
 /**
  * 舆情级别数据库服务实现
@@ -46,7 +47,7 @@ public class WarnLevelServiceImpl implements WarnLevelService {
 		sql.append(" on a.id = b.warn_level_id group by a.id");
 		return DBHandle.query(sql.toString(), new Object[0], WarnLevel.class);
 	}
-
+	
 	/**
 	 * 添加舆情级别
 	 * 
@@ -54,7 +55,12 @@ public class WarnLevelServiceImpl implements WarnLevelService {
 	 * @return
 	 */
 	@Override
-    public boolean add(WarnLevel warn) throws DBException {
+    public boolean add(CrudEntity crud) throws DBException {
+		if(!(crud instanceof WarnLevel)){
+			throw new DBException();
+		}
+		WarnLevel warn = (WarnLevel) crud;
+		
 	    StringBuffer sql = new StringBuffer();
 	    sql.append("insert into ").append(WarnLevel.DB_NAME);
 	    sql.append(" (`name`,`min_rate`,`max_rate`,`description`,`send_msg`,`need_handle`,`msg_tpl`) ");
@@ -76,7 +82,12 @@ public class WarnLevelServiceImpl implements WarnLevelService {
 	 * @return
 	 */
 	@Override
-    public boolean update(WarnLevel warn) throws DBException {
+    public boolean update(CrudEntity crud) throws DBException {
+		if(!(crud instanceof WarnLevel)){
+			throw new DBException();
+		}
+		WarnLevel warn = (WarnLevel) crud;
+		
 	    StringBuffer sql = new StringBuffer();
 	    sql.append("update ").append(WarnLevel.DB_NAME);
 	    sql.append(" set name=?,min_rate=?,max_rate=?,description=?,send_msg=?,need_handle=?,msg_tpl=? ");
@@ -97,7 +108,12 @@ public class WarnLevelServiceImpl implements WarnLevelService {
 	 * @return
 	 */
 	@Override
-    public boolean delete(WarnLevel warn) throws DBException {
+    public boolean delete(CrudEntity crud) throws DBException {
+		if(!(crud instanceof WarnLevel)){
+			throw new DBException();
+		}
+		WarnLevel warn = (WarnLevel) crud;
+		
 	    StringBuffer sql = new StringBuffer();
 	    sql.append("delete from ").append(WarnLevel.DB_NAME);
 	    sql.append(" where id=?");
